% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_demo_token}
\alias{get_demo_token}
\title{Returns token for demonstration}
\usage{
get_demo_token()
}
\value{
A string with token
}
\description{
Returns token for demonstration
}
\examples{
get_demo_token()
}
