% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividends.R
\name{get_dividends}
\alias{get_dividends}
\title{Fetches dividend data from eodhd}
\usage{
get_dividends(
  ticker = "AAPL",
  exchange = "US",
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{ticker}{a company ticker (identificer) (e.g. AAPL)}

\item{exchange}{a exchange (e.g. US). You can find all tickers and exchanges from get_info()}

\item{cache_folder}{A local directory to store cache files. By default, it uses a temporary path, meaning that the cache system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply
point to a local directory in your file system.}

\item{check_quota}{A flag (TRUE/FALSE) for wheter to check the quote from api or not (a small trade of execution time)}
}
\value{
A dataframe with dividend information
}
\description{
Fetches dividend data from eodhd
}
\examples{
\dontrun{
 df_div <- get_dividends(ticker = "AAPL", exchange = "US")
}
}
