% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{get_earnings}
\alias{get_earnings}
\title{Retrieves the earnings calendar}
\usage{
get_earnings(
  first_date = Sys.Date(),
  last_date = Sys.Date() + 30,
  symbols = NULL,
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{first_date}{the first date of the window. Default is today.}

\item{last_date}{the last date of the window. Default is in one month.}

\item{symbols}{an optional character vector of tickers with exchange
(e.g. c("AAPL.US", "MSFT.US")). Default (NULL) returns every company.}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with the earnings calendar
}
\description{
Queries \url{https://eodhd.com/financial-apis/calendar-upcoming-earnings-ipos-and-splits}
and returns reported and upcoming earnings, with estimate, actual value and
surprise.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_earnings <- get_earnings(symbols = "AAPL.US")
}
}
