% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{get_earnings_trends}
\alias{get_earnings_trends}
\title{Retrieves analyst estimates and their revisions}
\usage{
get_earnings_trends(
  symbols = "AAPL.US",
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{symbols}{a character vector of tickers with exchange
(e.g. c("AAPL.US", "MSFT.US"))}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with the analyst estimates
}
\description{
Queries the earning trends endpoint
\url{https://eodhd.com/financial-apis/calendar-upcoming-earnings-ipos-and-splits},
which returns, for each period, the average analyst estimate for earnings and
revenue, together with the revisions of the last months.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_trends <- get_earnings_trends("AAPL.US")
}
}
