% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_exchange_details}
\alias{get_exchange_details}
\title{Retrieves the details of an exchange}
\usage{
get_exchange_details(
  exchange = "US",
  first_date = NULL,
  last_date = NULL,
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{exchange}{A exchange symbol (e.g. US). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}. Be aware that, for US companies, the exchange symbols is simply "US"}

\item{first_date}{optional first date of the holiday window}

\item{last_date}{optional last date of the holiday window}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A list with the exchange details (including dataframes of trading
hours and holidays)
}
\description{
Queries \url{https://eodhd.com/financial-apis/exchanges-api-list-of-tickers-and-trading-hours}
and returns trading hours, holidays and the number of active tickers of an
exchange.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
l_details <- get_exchange_details("US")
}
}
