% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchanges.R
\name{get_exchanges}
\alias{get_exchanges}
\title{Fetches exchange list from api}
\usage{
get_exchanges()
}
\value{
a dataframe with information about available exchanges
}
\description{
Fetches exchange list from api
}
\examples{

# you need a valid token (not test) for this to work
\dontrun{
df_exc <- get_exchanges()
}


}
