% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro-rates.R
\name{get_funding_stress_spreads}
\alias{get_funding_stress_spreads}
\title{Retrieves the funding stress spreads}
\usage{
get_funding_stress_spreads(
  code = NULL,
  first_date = NULL,
  last_date = NULL,
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{code}{an optional code of the spread}

\item{first_date}{an optional first date of the window}

\item{last_date}{an optional last date of the window}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with the funding stress spreads
}
\description{
Queries the funding stress endpoint of eodhd, with the spreads used to monitor
the stress of the money market.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_spreads <- get_funding_stress_spreads()
}
}
