% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes.R
\name{get_historical_market_cap}
\alias{get_historical_market_cap}
\title{Retrieves the historical market capitalisation of a company}
\usage{
get_historical_market_cap(
  ticker = "AAPL",
  exchange = "US",
  first_date = Sys.Date() - 365,
  last_date = Sys.Date(),
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{ticker}{A company ticker (e.g. AAPL). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}.}

\item{exchange}{A exchange symbol (e.g. US). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}. Be aware that, for US companies, the exchange symbols is simply "US"}

\item{first_date}{the first date of the window. Default is one year ago.}

\item{last_date}{the last date of the window. Default is today.}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with columns date and value
}
\description{
Queries \url{https://eodhd.com/financial-apis/stock-etfs-fundamental-data-feeds}
(historical market capitalisation), which returns the daily market cap of a
company. History is limited to the last year for most subscriptions.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_cap <- get_historical_market_cap("AAPL", "US")
}
}
