% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{get_index_composition}
\alias{get_index_composition}
\title{Get Index Composition}
\usage{
get_index_composition(index, cache_folder = get_default_cache())
}
\arguments{
\item{index}{index to find composition (e.g. "GSPC.INDX")}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}
}
\value{
a list with three elements: info, current_components, historical_components
}
\description{
Retrieves the current and historical components of a financial index
from the EODHD API. Uses a caching system to avoid redundant API calls.
}
\examples{

\dontrun{
l_out <- get_index_composition("GSPC.INDX")
str(l_out)
}
}
