% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{get_index_list}
\alias{get_index_list}
\title{Retrieves the list of indices with composition data}
\usage{
get_index_list(cache_folder = get_default_cache(), check_quota = TRUE)
}
\arguments{
\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with the available indices
}
\description{
Queries the S&P Global data of the eodhd marketplace and returns the indices
that have components available for \code{\link[=get_index_composition]{get_index_composition()}}.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_indices <- get_index_list()
}
}
