% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday.R
\name{get_intraday}
\alias{get_intraday}
\title{Retrieves intraday data for a given ticker and exchange}
\usage{
get_intraday(
  ticker = "AAPL",
  exchange = "US",
  frequency = "5m",
  first_date = Sys.Date() - 7,
  last_date = Sys.Date(),
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{ticker}{A company ticker (e.g. AAPL). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}.}

\item{exchange}{A exchange symbol (e.g. US). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}. Be aware that, for US companies, the exchange symbols is simply "US"}

\item{frequency}{The frequency of the intraday data. Available options: "1m"= 1 minute, "5m" = 5 minutes, "1h" = 1 hour}

\item{first_date}{the first date of the query window. The function will keep querying the api
until this date is reached. Default is previous week.}

\item{last_date}{the last date of the query window. Default is today.}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with intraday prices (datetime, open, high, low, close, volume)
}
\description{
This function will query the intraday endpoint of eodhd
\url{https://eodhd.com/financial-apis/intraday-historical-data-api} and return all
bars available within a user supplied time period. Since the api limits how far
back a single query can reach, the function walks backwards in time, querying
consecutive windows until the first date is covered.
}
\examples{
\dontrun{
set_token(get_demo_token())
df_intraday <- get_intraday(ticker = "AAPL", exchange = "US")
}
}
