% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{get_news_word_weights}
\alias{get_news_word_weights}
\title{Retrieves the most relevant words of the news of a ticker}
\usage{
get_news_word_weights(
  ticker = "AAPL",
  exchange = "US",
  first_date = NULL,
  last_date = NULL,
  limit = 50,
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{ticker}{A company ticker (e.g. AAPL). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}.}

\item{exchange}{A exchange symbol (e.g. US). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}. Be aware that, for US companies, the exchange symbols is simply "US"}

\item{first_date}{an optional first date of the window}

\item{last_date}{an optional last date of the window}

\item{limit}{maximum number of words (default 50)}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with columns word and weight
}
\description{
Queries the news word weights endpoint of eodhd, which returns the keywords of
the news of a ticker with their weight.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_words <- get_news_word_weights("AAPL", "US")
}
}
