% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanctions.R
\name{get_sanctions_sources}
\alias{get_sanctions_sources}
\title{Retrieves the sources of the sanction lists}
\usage{
get_sanctions_sources(cache_folder = get_default_cache(), check_quota = TRUE)
}
\arguments{
\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with the sources
}
\description{
Lists the sanctioning bodies behind the consolidated data, with the date of
their last update.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_sources <- get_sanctions_sources()
}
}
