% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanctions.R
\name{get_sanctions_vessels}
\alias{get_sanctions_vessels}
\title{Retrieves sanctioned vessels}
\usage{
get_sanctions_vessels(
  source = NULL,
  imo = NULL,
  flag = NULL,
  vessel_type = NULL,
  query = NULL,
  program = NULL,
  limit = 100,
  offset = 0,
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{source}{an optional source of the list (see \code{\link[=get_sanctions_sources]{get_sanctions_sources()}})}

\item{imo}{an optional IMO number of the vessel}

\item{flag}{an optional flag (country) of the vessel}

\item{vessel_type}{an optional type of vessel}

\item{query}{an optional free text search}

\item{program}{an optional sanction program (see \code{\link[=get_sanctions_programs]{get_sanctions_programs()}})}

\item{limit}{maximum number of rows (default 100)}

\item{offset}{pagination offset (default 0)}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with the sanctioned vessels
}
\description{
Queries the sanctions endpoints of eodhd for the vessels of the consolidated
lists, searchable by IMO number and flag.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_vessels <- get_sanctions_vessels(limit = 10)
}
}
