% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{get_splits_calendar}
\alias{get_splits_calendar}
\title{Retrieves the splits calendar}
\usage{
get_splits_calendar(
  first_date = Sys.Date(),
  last_date = Sys.Date() + 30,
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{first_date}{the first date of the window. Default is today.}

\item{last_date}{the last date of the window. Default is in one month.}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with the splits calendar
}
\description{
Queries \url{https://eodhd.com/financial-apis/calendar-upcoming-earnings-ipos-and-splits}
and returns the stock splits of a period, including the upcoming ones.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_splits <- get_splits_calendar()
}
}
