% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Retrieves a list of tickers for a particular exchange}
\usage{
get_tickers(exchange = "US")
}
\arguments{
\item{exchange}{a exchange (e.g. US). You can find all tickers and exchanges from get_info()}
}
\value{
A dataframe with a list of tickers
}
\description{
Retrieves a list of tickers for a particular exchange
}
\examples{
\dontrun{
df_tickers <- get_tickers("US")
}
}
