% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{parse_financials}
\alias{parse_financials}
\title{Parses financial data from the API}
\usage{
parse_financials(l_out, type_table = "wide")
}
\arguments{
\item{l_out}{List from get_fundamental()}
}
\description{
Parses financial data from the API
}
\examples{
# no example

}
