% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_token}
\alias{set_token}
\title{Sets up authentication for eodhd}
\usage{
set_token(token = get_demo_token())
}
\arguments{
\item{token}{the token from eodhd. The default value is a demo token "demo", which allows for partial access to the data}
}
\value{
Nothing
}
\description{
Uses the token from \url{https://eodhd.com/cp/dashboard} to authenticate your R session. You can find your own eodhd token from the website.
Althernatively, a demo token is also available for testing purposes.
}
\examples{
set_token()
}
