% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchanges.R
\name{get_exchanges}
\alias{get_exchanges}
\title{Retriees the list of exchanges available}
\usage{
get_exchanges()
}
\value{
a dataframe with information about available exchanges
}
\description{
Retriees the list of exchanges available
}
\examples{

# you need a valid token (not test) for this to work
\dontrun{
set_token(YOUR_VALID_TOKEN)
df_exc <- get_exchanges()
}

}
