% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_options_underlyings}
\alias{get_options_underlyings}
\title{Retrieves the symbols with options data}
\usage{
get_options_underlyings(
  limit = 1000,
  offset = 0,
  cache_folder = get_default_cache(),
  check_quota = TRUE
)
}
\arguments{
\item{limit}{maximum number of rows (default 1000)}

\item{offset}{pagination offset (default 0)}

\item{cache_folder}{A local directory to store cache files. By default, all functions use a temporary path, meaning that the caching system
is session persistent (it will remove all files when you exit your R session). If you want a persistent caching system, simply point argument
cache_folder to a local directory in your filesystem. Be aware, however, that a persistent cache will not refresh your data for new api queries.}

\item{check_quota}{A flag (TRUE/FALSE) for whether to check the current quota status from the api. This option implies a small cost of execution
time. If you need speed, just set it to FALSE.}
}
\value{
A dataframe with one column, underlying_symbol
}
\description{
Queries the Unicorn Bay options data of the eodhd marketplace
\url{https://eodhd.com/marketplace/unicornbay/options}, and returns the
underlying symbols that have contracts available.
}
\examples{
\dontrun{
set_token("YOUR_VALID_TOKEN")
df_underlyings <- get_options_underlyings()
}
}
