% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Retrieves a list of tickers for a particular exchange}
\usage{
get_tickers(exchange = "US")
}
\arguments{
\item{exchange}{A exchange symbol (e.g. US). You can find all tickers for a particular exchange with \code{\link[=get_tickers]{get_tickers()}}.}
}
\value{
A dataframe with a list of tickers
}
\description{
Retrieves a list of tickers for a particular exchange
}
\examples{
\dontrun{ # requires a subscription (paid) token
df_tickers <- get_tickers("US")
}
}
