% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{parse_financials}
\alias{parse_financials}
\title{Parses financial data from the API}
\usage{
parse_financials(l_out, type_table = "wide")
}
\arguments{
\item{l_out}{List from get_fundamental()}

\item{type_table}{Type of table in output ("wide" or "long"). A wide table is a typical Excel column-oriented table where each columns is a data/year.
A long type of table row-oriented, where each each point of new information is a row of the table.}
}
\description{
Parses financial data from the API
}
\examples{
# no example

}
